Fix more wrong output types
.. compare to dhfun1

+++ +dh/getwaveletsampleperiod.m
@@ -15,4 +15,4 @@
 function samper = getwaveletsampleperiod(fid, blkid)
 
 filename = get_filename(fid);
-samper = h5readatt(filename, "/WAVELET" + blkid, 'SamplePeriod');
+samper = double(h5readatt(filename, "/WAVELET" + blkid, 'SamplePeriod'));
